+++ packages/console-types/src/index.ts
@@ -90,6 +90,7 @@ import {
   services,
   swaggerPath,
   url,
+  environmentVariables,
 } from './types/services'
 import { decorators, postDecorator, preDecorator } from './types/decorators'
 import { endpoint, endpoints, path } from './types/endpoints'
@@ -138,6 +139,7 @@ export {
   dockerImage,
   endpoint,
   endpoints,
+  environmentVariables as serviceEnvironmentVariables,
   listeners,
   environment,
   externalService,